Tweaks to Peters skiss to try out initialize manually instead. We now generate some kind of content for _static_kernel_objects_begin and gen_kobject_list.py complains about something else than not finding it.
Load$$section-name does not work for initialize manually. So refer to section-name_init instead.
Also, fix broken _end symbol for groups, where we used ${name_clean} instead of ${last_section_name}

Fix so that we generate "do not initalize" for noinit blocks in icf files.

Ensure that we put keep on the _init-blocks
Fix a bunch of typos and improve some comments

Add output of keep{} for the blocks too. otherwise we dont REALLY keep unless we have symbols....