cmake: linker generator: Handle input from prev linker pass
Add methods for the linker-script generator to pick up information
from previus linkerpasses:
 * zephyr_linker_include_generated(FILE f.cmake [PASS X]) cmake-includes
   f.cmake in the relevant PASSes. This allows scripts to generate any
   content (e.g. sections or section configs) as if it was generated by
   the main cmake machiery. This is intended to cover the case of
   ld-script snippets generated by e.g. gen_app_partitions.py
 * @FOO[,undef:v]@ in an value or expression in the zephyr_linker_*
   inputs. This looks for a defintion of FOO (se below), and uses its
   value if found. If FOO is not defined, the whole @@ thing is left
   untouched, unless ,undef: is used, in which case v is used instead.
 * zephyr_linker_include_generated(FILE f.h [PASS x]) greps for f.h
   for #define FOO value to be accessed as @FOO@
 * zephyr_linker_include_var(VAR FOO VALUE V) allows the main
   cmake-script to set a @FOO@ variable.

Note that the #define support is VERY basic, and does not use a proper
pre-processor. It works for the current use of files generated by e.g.
gen_kobject_placeholders.py but is not a general solution.

Use the new mechanisms to to handle ${DEVICE_API_LINKER_SECTIONS_CMAKE},
the kobject-prebuilt-*.h files and the apt_smem defintions.