cmake: linker generator: userspace: Fix generator for APP_SMEM
Tweak gen_app_partitions.py to be able to generate cmake ld-script
generator output as well as ld-script fragments.

This removes the hardcoded partition in linker.cmake and replaces
it with a group to contain the generated partitions.

Setup generator variables for alignment of APP_SMEM. Note that this does
not yet handle MPU_ALIGN which depends on the size of the section...

Use said variables in the sections generated by gen_app_partitions.py

Move APP_SMEM_GROUP from DATA_REGION to RAM_REGION since it has its
own copy-stuff in z_data_copy()